+++ src/database/migrations/1788591482605-CreateUploadTable.ts
@@ -19,6 +19,10 @@ export class CreateUploadTable1788591482605 implements MigrationInterface {
   }
 
   public async down(queryRunner: QueryRunner): Promise<void> {
+    const env = process.env.NODE_ENV ? process.env.NODE_ENV.toLowerCase() : '';
+    if (!['development', 'dev', 'local'].includes(env)) {
+      throw new Error('Migrations can be reverted in development, dev or local environment');
+    }
     await queryRunner.query(`DROP INDEX \`IDX_uploads_file_path\` ON \`uploads\``);
     await queryRunner.query(`DROP TABLE \`uploads\``);
   }
