+++ src/database/migrations/1789110998856-CreateRefreshTokensTable.ts
@@ -27,6 +27,10 @@ export class CreateRefreshTokensTable1789110998856 implements MigrationInterface
   }
 
   public async down(queryRunner: QueryRunner): Promise<void> {
+    const env = process.env.NODE_ENV ? process.env.NODE_ENV.toLowerCase() : '';
+    if (!['development', 'dev', 'local'].includes(env)) {
+      throw new Error('Migrations can be reverted in development, dev or local environment');
+    }
     await queryRunner.query(`
       DROP TABLE IF EXISTS \`refresh_tokens\`
     `);
